fix: Ensure adornment appears after attribute changes
PT Story: https://www.pivotaltracker.com/story/show/181940090

This fixes a bug where starting from a configuration where the x attribute type is numeric and the y attribute type is categorical, to a configuration where the x attribute is numeric and the y attribute is numeric, and then switching back to the previous configuration, the adornment's line would not appear in the plot even if its checkbox was checked and a value provided.

Also included are fixes to correct some styling issues, and ensure that a formula value will use the correct case data when there are multiple subplots.