fix: SD hover tip values are wrong (PT-186447543)
PT Story: https://www.pivotaltracker.com/story/show/186447543

There was a small issue with the previous work on this story in that the values being shown in the hover tips could sometimes be very slightly off when compared to V2 (e.g. 15.2 versus 15.1). To get values that match V2's exactly, this changes `adornmentSpecs` to use the unformatted `value` instead of the formatted `displayValue` in computing`rangeValue` which is used for the SD tip text.