fix: Percent adornment should show subcategories with split plots (PT-186789144)
PT Story: https://www.pivotaltracker.com/story/show/186789144

This begins to fix the problem, but I want to check some things before going further. I'm not sure we want to always be showing the percent options whenever v2 does. There was some discussion about that a while ago that I'm not remembering clearly. I also suspect the bug described in https://www.pivotaltracker.com/story/show/186781301 may be causing issues with the percent calculations, though that may not be true. In any case, the percents for Cell and Row are not yet being calculated properly when there are three categorical attributes present.