[#187799278] Feature: A slider can be configured to display date-time values
This feature is required by the DST project so that users can animate through dates and times, not just numerically, to drive formulas that change attribute values and displays of data.
Prior to this PR the only way for a user to change the min and max of a slider is by dragging. This won't be practical with a date-time axis, so we add widgetry in the inspector panel with which the user can set a min and max.
* In axis-model.ts we need to be able set minimum and maximum of `BaseNumericAxisModel` separately.
* Note that a date axis usually requires two lines. In this PR we are making use of surplus space below the value, but this is an interim solution. The longer term solution is to change the height of the slider component as needed. We will write a PT story about this.
* Various improvements to the inspector palettes
* Various improvements to the animation routine
* Modified multi-scale.ts to have an observable property `numericDomain` that mirrors the numericScale.domain so that computed properties, especially `resolution` will be recomputed as needed. This is necessary because numericScale.domain as a property of a D3 scale is *not* observable.
* There is a bit of a hack in use-sub-axis.ts, line 239 where we special case categorical axes to prevent a crash that occurs on redo of assigning a categorical attribute. Considerable investigation didn't turn up a better solution.
* Skipping most of the cypress tests for Talya to reimplement.