if interactiveApi url param, then hide cfm menu
The interactieApi url param is added to the URL when a activity player embedded share link is copied
from the CFM share menu.
The CODAP v2 code also checked for older url params `lara` and `launchFromLara`. There are no interactives
in the LARA authoring system which use those older params, so support for them is dropped.