send content unwrapped
The CFM expects the content to be unwrapped when we send it.
This was mostly working before because the CFM was fooled into think it had wrapped this content instead of us. So then it would unwrap it before saving because of the `wrapFileContent` setting. However the interactiveAPI provider was not doing this incorrect unwrapping when saving. Instead the CFM would unwrap the file when loading the file but because the CFM was expecting the file to be unwrapped to start with. Because of this the document would be marked dirty every time it was re-opened in the activity player.