update window title when the document name changes
This replaces the document title in the document state with a volatile title property. It also removes the filename from the properties. Both changes are intended to force the CFM to be the source of truth for the title.
This change of state should not require a migration of existing v3 documents because the filename/name was already managed by the CFM and just not updated in v3 correctly.

To keep the title in sync the filename needs to be processed when the CFM sends the openedFile, savedFile, and renamedFile events.