Input row has overlay but has the wrong width.
Bugs:
Input row drag and drop works but highlight and drop is off by 1 or 2 rows.

Dragging up beyond a certain row also captures the attribute dropzone instead of the row drop zone.

Users can drag input row to a different collection, but the drop is in the input row collection