async api handlers
Refactor the handling of api requests from interactives.
The handlers can now be asynchronous. If an array of api requests is sent by the interactive they will be run one after the other which each waiting for the previous one to finish.
Individual api requests will be run when they come in even if there is another api request or array of request waiting to finish.

The api request loop is changed from an autorun based observer to a reaction observer. This way properties accessed by each handler are not observed. We only want to run the loop when a new request comes in, not when a property changes that a handler accessed.