breakup some dependency trees
Add a IBaseDataDisplayModel for use by the legend code so it doesn't depend on the full DataDisplayModel which brings in a dependency on the Tile System.

Update the types and type checking of the GraphContentModelContext and MapContentModelContext to be safer. They will now throw a runtime error if the wrong context object is used.

Move the default data display model which was not really function and seemed to only be used for tests in the test that needed it: use-axis-test.tsx.

Split out axis-constants which used to be in graphing-types.