fix issues found by better typing
Turns out that useAxis is used by the slider component, and in that case there is no real DataDisplayModel.

The methods in the default DataDisplayModel that was removed was actually used by the Numeric axis model. So these methods were pushed into the numeric axis model to make it more clear.

The useSubAxis was assuming in some points the displayModel was a full display model. For example by calling applyModelChange. This call would have failed if it was reached within the slider component.

The AxisProviderContext was also providing a default fake axis provider. In this case the empty axis provider was never used (as far as I could tell), so I removed the default and had useAxisProviderContext throw an error to simplify some of the typing.

In axis-helper it was assuming the axisBounds was never undefined, but technically it could be, so now it throws an explicit error instead of some harder to find error which would have shown up before.