beginning of history service
The idea is that within CODAP the "app" history service will handle the applyModelChange. When the code is used outside of CODAP a no-op history service can be used.

Part of this work is also to type the INotification better. Instead of just using the iframePhone generic types, now the notifications are type specific to how CODAP is using them. The core of CODAP will work with ICoreNotifications. And then things like drag and drop support for iframes extend ICoreNotifications.  The rational is that the notification system isn't necessary specific to iframePhone or iframes at all. It is just a generic notification system which is then exposed to plugins via the iframe API.