break dependencies around the formula system
- move a couple formula adapter registrations to app.tsx
- add a IFormulaManager interface to be used by the TileEnvironment so the full FormulaManager doesn't need to be exposed to tile content models and their sub objects.
- move abstract base class FormulaManagerAdapter to its own file along with the IFormulaAdapterApi. This breaks the dependency on the BoundaryManager and GlobalValueManager
- createCodapDocument makes the assumption that the IFormulaManager returned by getFormulaManager is really a FormulaManager. Given that it basically configured all this in createDocumentModel it is safe to assume.
- update tests so Formula object an get the formulaManager from its MST environment.
- add 3 new methods to FormulaManager and IFormulaManager so that the Formula model does not directly depend on the formula system. Instead the Formula model uses the formulaManager to do its work. Unfortunately, this re-enforces the existing circular dependency between the Formula and FormulaManager. But at least now the circle is based on interfaces not actual class types.
- move ICase out of formula-types.ts since that class imported mathjs.

Note: this does not fix some existing circular dependencies. For example: Formula -> IFormulaManager -> IFormulaManagerAdapter -> IFormula
I don't have enough time to figure out how to fix that now. It is just types that are circular so it won't cause a problem at runtime.

Note 2: The formula adapter system could probably be simplified by having the adapters look up their content models via a new method on IFormulaAdapterApi. This Api knows the document so it could just search the document for the specified tile types and return their contentModels. This way content models don't have to add themselves to the adapters when they are added to the document. I didn't make this change because things were getting too large already.