The idea is that within CODAP the AppHistoryService will handle the applyModelChange. When the code is used outside of CODAP a no-op history service can be used.
Part of this work is also to type the INotification better. Instead of just using the iframePhone generic types, now the notifications are typed specific to how CODAP is using them. The core of CODAP will work with ICoreNotifications. And then things like drag and drop support for iframes extend ICoreNotifications. The rational is that the notification system isn't necessary specific to iframePhone or iframes at all. It is just a generic notification system which is then exposed to plugins via the iframe API.

Additionally this takes a new approach for the ITileEnvironment. It is now created by intersecting several individual environment interfaces. This makes it possible for services like the history service to depend on them in a way where the "key" of the service in the environment is typed in a single place while not making the history service dependent on the full ITileEnvironment.

When the tests are made part of the CODAP "library", instead of importing using AppHistoryService they'd use a no-op history service. Or if they are actually checking the history then a simplified history service could be used.

A variation of this PR could be have applyModelChange handle the case of not finding the history service. This would remove the need for most of the tests to setup MST environments with a historyService. However this will make it less obvious how to fix a new test that is expecting the history to work. So my preference is to fail fast and require the more explicit configuration.