split out v2 dataset conversion
data set notifications convert cases and attributes to v2 form
This conversion brings in several files and probably is unnecessary in a library.
So an adapter is added which can convert the attributes and cases.
A default adapter is configured which does no conversion.

Hopefully this won't cause too much confusion if someone is tracing through the code and trying to figure out why what they see in the plugin messages doesn't match what they expect.

This might also result in reduced code coverage. So then it might be useful to have some Jest tests that install the V2Adapter.