make cachedFnWithArgsFactory observable
This should not impact performance unless a lot of keys are used. I'd assume this cached function approach shouldn't be used when there are lots and lots of different keys.

I'd guess making this observable could eliminate several places where extra reactions are used around the dataConfigurationModel. I'm only aware of one specific case within the categorical legend which will be part of a different PR.