legend type based on the data configuration
The user can set an attribute type in the data configuration of the graph or map. The legend was ignoring this and just showing the type from the table.

Additionally the component wasn't observing the attribute type. So if the user changed the type in the table and they hadn't overridden it with a setting in the graph or map it wouldn't update.