add option to use quantized choropleth legend
many of the changes are just renaming methods from quantile to numericColor, choropleth, or bin

To support reactions the colorHash method was replaced with a legendColorDomain function. I'm not sure that is the best name.

Additionally to simplify the monitoring of category colors the hashing of colors in colorHash is replaced with a colorMap property. This colorMap is then used to speed up the the access of each category color. With the colorHash approach MobX was monitoring every category anyway, so it makes sense to expose this. It has the same downside as colorHash:  changing a single category color will cause all points to be re-rendered.