fix: switching plot type results in adornmentList API confusion (CODAP-342)
[CODAP-342](https://concord-consortium.atlassian.net/browse/CODAP-342)

This fixes an issue where adornments activated in one plot configuration, were still being listed for `adornmentList` requests after the plot had changed to a type that those adornments don't support. It also adds a listing for a "Percent" adornment. Even though Percent is part of Count, plugin developers may not be aware of that since the CODAP UI presents Count and Percent as totally separate options.