feat: count adornment get returns counts per region (CODAP-68) (#1873)
* feat: count adornment get returns counts per region (CODAP-68)

[CODAP-68](https://concord-consortium.atlassian.net/browse/CODAP-68)

When a Movable Value is present and divides a graph plot into multiple regions, we should return a count per region in the response to `get` requests for the Count adornment.

These changes move most of the logic for computing region counts from the Count adornment component to the model so it can be shared with the API handler code. One thing I'm not so sure of is the new `createScale` function in the handler file. I couldn't find another way to get scale information in the handler functions, but I may be overlooking something.

---------

Co-authored-by: Kirk Swenson <kswenson@users.noreply.github.com>