CODAP-605 CODAP-605 Date Slider API
This also adds support for updating sliders, and adds validation of many of the properties sent during create or update.

A question about the new fields is the use of strings for values. The existing fields of animationDirection and animationMode use indexes. The strings are consistent with the underlying modeling and in other parts of the CODAP api string are used.