[#CODAP-702] Bug fix: Using API, create of parent case returns caseID of child
The problem arose as part of getting the Markov game plugin to work in V3. The existing response to creating a parent case in a 2-level dataset was returning the caseID of the *child* case created in association with the parent case instead of the caseID of the parent case which is required by Markov.

This fix impinges on performance critical code in which a plugin creates cases. I am not at all confident that the fix is as performant as it can/should be!

* In the `create` method of case-handler.ts we make sure to return the case id whose collection is the one specified in the request. We also find its corresponding itemID because that appears to be part of the contract.