CODAP-740 CODAP using extracted formula models
This hasn't been tested for real yet, but the jest tests and build are passing.

It required several changes to the CODAP code so that it can pass datasets and the global manager to the formulas library. This could probably be avoided by generics in the formulas library but I've run out of time so decided to go with brute force.

The formula library tests are left in CODAP because they use the full CODAP document to test, so it would take a lot of mocking to move them into formula library.

Changes:
- new properties are added to the formula library dataset to reduce some of the casting needed between the formula dataset and the codap dataset.
- function-utils test was moved into formulas library it didn't have any codap dependencies. Doing this required updating the formulas/package.json so Jest could handle importing the various libraries used by the test, including the utilities library.
- several files were updated to cast CODAP types to formulas library types, these colud be fixed with generics. They are documented in the formulas/todos.md
- The existing CaseList type was used:
  - data-display-filter-formula-adapter.ts
- multiple functions tests were updated to use the functions from the formulas library. Some functions tests did not directly use the function module they were testing instead just using string based evaluation.
- the function modules were removed from CODAP because they weren't needed anymore.
- several files were left in CODAP but they just export all of the properties from their counterpart in the formulas library. The intention is to reduce the amount of changes needed in CODAP so this PR has fewer conflicts.