CODAP-740 disable formula editor tabs, fix formula editor height
the previous approach of passing a style for setting the height didn't seem to really be working. There is an element inside of the element that gets the style which then has a different height. Passing the height property to CodeMirror and removing this style fixes that problem.

The tabs are disabled so keyboard users can tab through the editor without getting stuck in it.