CODAP-57 Make tile transitionComplete a volatile property
Additionally, transitionComplete is now set to true immediately if the tile is not animated into its final size.

This change required changing the way the map rescaling code worked. The is a new map.md file describing how it should work. The problematic case was how the code was rescaling the map to fit the points of a dataset. This was relying on transitionComplete being a serialized property. transitionComplete was used to prevent a map created with a center and zoom from being rescaled. It was also used to prevent a map restored from serialized state from being rescaled again which would override any changes made by the user.

The fix for these issues is to have the code that automatically calls rescale to make sure the map doesn't already have a set zoom before rescaling. zoom is a serialized property and after a successful rescale it will be set.

Additionally the code which automatically rescales was moved into useMapModel in a MobX `when`.

Other Fixes:
- when a map is created by the plugin api in a document with a dataset with latitude and longitude attributes it will now automatically fit this data like CODAPv2 does. That is unless the plugin sets a zoom value for the new map.
- when a dataset has a single mappable point the map would zoom in too far and not have any map tiles to show at this zoom level. This was fixed by adding a maxZoom setting of 12. The value of 12 is kind of arbitrary. At this level the user can probably still see some cities or roads that might provide visual context. But that depends on where in the world the point is located.
- when the noComponentAnimation url param is used the text tile used to not automatically get focused and also not be focused when the tile was selected.
- when the text tile was created by the api it would also not get automatically focused and also not be focused when the tile was selected.

Other minor changes:
- update usage of useTileModelContext to be consistent
- when tile is created you can no longer specify transitionComplete. This required transitionComplete to be part of the serialized properties, and was only used by the map code.