CODAP-900 use full document applySnapshot in di doc update
This approach fixes some problems. It might cause other problems. It hasn't been tested with many dataset cases yet.

It is a little fragile because when the shared model map is round tripped through v2, the keys of the shared model entries are lost. So there has to be logic to find the matching shared model entry in the incoming snapshot and update its key to match the existing key.

The changes to the shared models are to add type guards to make the code more type safe.