fix: Remove file extension on iPad saves [CODAP-303]
This fixes the issue where ".txt" was being appended to files saved from iPad.

This was due to the way iOS handles file downloads, and the fix involves
updating the Cloud File Manager configuration to set the new
localFileMimeType config option to 'application/octet-stream' which prevents the
addition of file extensions in the CFM.

The @concord-consortium/cloud-file-manager package was updated to the latest
version which ensures that the localFileMimeType is used and file extension
handling is correct when saving files locally.