feat: Implement import url functionality [CODAP-84]
This commit adds the ability to import a CODAP document from a URL via the CFM in v3.

The new import code shares a code path with the drag-and-drop import code.

This also fixes the css in the Import URL dialog to make the input field full width and match other input padding.