feat: Implement geojson file/url drop handling [CODAP-890]
This commit adds support for handling geojson files and URLs dropped into the application
as well as support for geojson URLs imported via the "Import from URL" dialog.

When a geojson file is dropped, it initiates a generic import with the appropriate content type.

This also includes a generic import utility function to streamline the import process for
other file types in the future.  The current CSV import handling remains unchanged.